Added frame, animation, direction, spritesheet, and sprite dimension/offset values to Mobile, as well as a non-abstract definition of draw() to use them. Made small changes to Bullet and Sprite, added javadocs.

+++ src/SpriteComparitor.java
@@ -4,6 +4,6 @@ public class SpriteComparitor implements Comparator<Sprite>{
 
     @Override
 	public int compare(Sprite x, Sprite y) {
-		return Integer.compare(x.getLayer(), y.getLayer());
+		return Double.compare(x.getLayer(), y.getLayer());
 	}
 }
